added support for protocol V2 and V4 updated tests

+++ tests/IntegrationTestCase.php
@@ -3,6 +3,7 @@
 namespace GraphAware\Bolt\Tests;
 
 use GraphAware\Bolt\Configuration;
+use GraphAware\Bolt\Driver;
 use GraphAware\Bolt\GraphDatabase;
 
 class IntegrationTestCase extends \PHPUnit_Framework_TestCase
@@ -17,11 +18,13 @@ class IntegrationTestCase extends \PHPUnit_Framework_TestCase
      */
     protected function setUp()
     {
-        $this->driver = GraphDatabase::driver(
+        $version = getenv('BOLT_VERSION') ? getenv('BOLT_VERSION') : 0;
+        $this->driver = new Driver(
             $this->getBoltUrl(),
             getenv('NEO4J_USER') ?
                 Configuration::create()->withCredentials(getenv('NEO4J_USER'), getenv('NEO4J_PASSWORD'))
-                : Configuration::create()
+                : Configuration::create(),
+            (int)$version
         );
     }
 
@@ -69,9 +72,19 @@ class IntegrationTestCase extends \PHPUnit_Framework_TestCase
         $indexRecords = $this->getSession()->run('CALL db.indexes()');
         $drops = array_map(function ($record){
             $name = $record->get('name');
+            // older neo4j version
+            if(!$name){
+                if ($record->get('type') == 'node_label_property') {
+                    return 'DROP '.$record->get('description');
+                }
+                $label = $record->get('label');
+                $property = $record->get('properties')[0];
+                return "DROP CONSTRAINT ON (n:$label) ASSERT n.$property IS UNIQUE";
+            }
             if (strpos($name, 'index') !== false){
                 return "DROP INDEX $name";
             }
+
             return "DROP CONSTRAINT $name";
         }, $indexRecords->getRecords());
         foreach ($drops as $drop) {
